ANDROID: mm: add vendor hook in isolate_freepages()
By this vh, so that we can skip if this page needs to be
treated specially, l(CR) as cma.

In the Android kernel, the use of cma is restricted, and filecache
cannot use cma.But during the memory compaction process, filecache
may be migrated to cma pool, so a judgment needs to be added here
to restrict filecache from entering cma.

Bug: 309371168
Change-Id: I3ec29bdf5f7b6ac4c7af0a317aa41ad77b71444d
Signed-off-by: Qinglin Li <qinglin.li@amlogic.com>