usb-dwc3 Fix invalid charger type detection for float
Qcom kernel-4.19/5.4 have below fix which not porting to kernel new version
Float/unknow charger input current be override by usb set setting value.
Re-adding below logic to new kernel version for such case which enhance the
charging capability compatible.
The charger driver on certain platform depends on the USB to differentiate
between a Standard Downstream Port and a Float charger, so that the case of
slow insertionis taken care of. The USB driver has to set the current to
-ETIMEDOUT for float and a valid value (100mA, 500mA, 900mA)for SDP so
that the charger can differentiate between the two.Add the support for the
same by checking the link state in a second in the case the charger detected
is Float to differentiate between Standard Downstream Portand a Float charger.

Change-Id: I4645e2dfb87f7ff4c4b0e5f51b731b8e2e913f15
Signed-off-by: xuwei9 <xuwei9@lenovo.com>
Reviewed-on: https://gerrit.mot.com/2893144
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Xiangpo Zhao <zhaoxp3@motorola.com>
Reviewed-by: Wang Wang <wangwang1@lenovo.com>
Reviewed-by: Zonghua Liu <a17671@motorola.com>
Submit-Approved: Jira Key