qcom wdog: delay wdog bark time interval
System will sometime crash due to the qcom watchdog, when the
system overload is very busy, it's easy to not response dog pet
action and cause the system reboot.

So we change watchdog bark time to 16 sec to reduce dog bit chance
when the system is a little busy.

Change-Id: If92c7bcb238c9dcba4ec1ab03c4c9bfb450b7996
Signed-off-by: qianlong <qianl2@motorola.com>
Reviewed-on: https://gerrit.mot.com/3041779
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Zhenxin Xi <xizx@motorola.com>
Submit-Approved: Jira Key
Reviewed-on: https://gerrit.mot.com/3414604