adsprpc : Handle UAF scenario in put_args
Currently, the DSP updates header buffers with unused DMA handle fds.
In the put_args section, if any DMA handle FDs are present in the
header buffer, the corresponding map is freed. However, since the
header buffer is exposed to users in unsigned PD, users can update
invalid FDs. If this invalid FD matches with any FD that is already
in use, it could lead to a use-after-free (UAF) vulnerability.
As a solution,add DMA handle references for DMA FDs, and the map for
the FD will be freed only when a reference is found.

Acked-by: quic_anane <quic_anane@quicinc.com>
Signed-off-by: Ansa Ahmed <quic_ansa@quicinc.com>

Mot-CRs-fixed: (CR)
CVE- fixed: CVE-2024-43047
CRs Fixed: QC-CR#3883647
Signed-off-by: Sipra Patel <siprap@motorola.com>

Change-Id: Id625558542a976746f34656ab9076c9882b06d15
Reviewed-on: https://gerrit.mot.com/3086132
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Varun Shrivastava <varunshrivastava@motorola.com>
Reviewed-by: Xiangpo Zhao <zhaoxp3@motorola.com>
Submit-Approved: Jira Key