arm64: dts: rockchip: Fix LED triggers on rk3308-roc-cc
[ Upstream commit 3a53a7187f41ec3db12cf4c2cb0db4ba87c2f3a1 ]

There are two LEDs on the board, power and user events.
Currently both are assigned undocumented IR(-remote)
triggers that are probably only part of the vendor-kernel.

To make dtbs check happier, assign the power-led to a generic
default-on trigger and the user led to the documented rc-feedback
trigger that should mostly match its current usage.

Fixes: 4403e1237be3 ("arm64: dts: rockchip: Add devicetree for board roc-rk3308-cc")
Cc: Perry Yan <Perry.yan@rock-chips.com>
Reviewed-by: Dragan Simic <dsimic@manjaro.org>
Signed-off-by: He(CR) Stuebner <he(CR)@sntech.de>
Link: https://lore.kernel.org/r/20241008203940.2573684-8-he(CR)@sntech.de
Signed-off-by: He(CR) Stuebner <he(CR)@sntech.de>
Signed-off-by: Sasha Levin <sashal@kernel.org>