usb: dwc3: msm: Handle absence of disconnect event during plug-out
In some situations, it is observed that during a faulty connect
where enumeration has failed, the controller goes into a bad state
where it is not able to generate any events. As a result when a plug
out happens in this scenario, the disconnect event isn't generated
by the controller and after the 2 second timeout, we exit the start
peripheral call setting WAIT_FOR_LPM bit. But since the connected
flag was never set to false, the idle/suspend routines for core
fail and the wrapper's suspend routine too is not called. This
instability causes further Plug-In from enumerating the device and
further plug-outs to enter LPM.

As per the debug analysis done so far, this is due to either AXI
response hang or the TBT command (internal)/response hang. While
debug is in progress, SNPS suggested restarting the device mode.

Fix this by forcing disconnect during faulty plug-out and fix the
gadget connect during next plug-in. Also since the user is free to
do a composition switch after faulty plug-out and before next plug
in, ensure to check whether udc is binded or not to avoid duplicate
calls to gadget_connect.

Change-Id: Ica53db16dc6d51bf652679ecfa967347c6c5f384
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>
Signed-off-by: Pratham Pratap <quic_ppratap@quicinc.com>
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>
Reviewed-on: https://gerrit.mot.com/3685341
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Xu Zhang <zhangxu21@lenovo.com>
Submit-Approved: Xu Zhang <zhangxu21@lenovo.com>
Tested-by: Xu Zhang <zhangxu21@lenovo.com>