ANDROID: KVM: arm64: Fix THPs reclaim with ballooning
A ppage content can't be read without either that ppage out of the
pinned_page tree or without the mmu_lock taken. Fix potential race
between vCPUS to decrease ppage->pins and free the page.

Also the memory accounting was wrong. Once the huge-page is freed, we
need to account for that entire size.

Bug: 278749606
Bug: 278011447
Bug: 400307522
Fixes: e602818486b6 ("ANDROID: KVM: arm64: Huge page support for pkvm_pinned_page")
Change-Id: I1a9f29ce9e63d707f59129aafd78e3d1d858614f
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>